+++ Resources/Prototypes/Entities/Mobs/Player/silicon_base.yml
@@ -229,7 +229,7 @@
           type: InstrumentBoundUserInterface
           requireInputValidation: false
         enum.RadialSelectorUiKey.Key:
-          type: RadialSelectorMenuBUI
+          type: TrackedRadialSelectorMenuBUI # WD EDIT: RadialSelectorMenuBUI -> TrackedRadialSelectorMenuBUI
         enum.ListViewSelectorUiKey.Key:
           type: ListViewSelectorBUI
         enum.SurgeryUIKey.Key:
