+++ Resources/Prototypes/Entities/Objects/Materials/Sheets/other.yml
@@ -70,7 +70,7 @@
   - type: UserInterface
     interfaces:
       enum.RadialSelectorUiKey.Key:
-        type: RadialSelectorMenuBUI
+        type: AttachedRadialSelectorMenuBUI # WD EDIT: RadialSelectorMenuBUI -> AttachedRadialSelectorMenuBUI
   - type: ActivatableUI
     key: enum.RadialSelectorUiKey.Key
     inHandsOnly: true
@@ -137,7 +137,7 @@
   - type: UserInterface
     interfaces:
       enum.RadialSelectorUiKey.Key:
-        type: RadialSelectorMenuBUI
+        type: AttachedRadialSelectorMenuBUI # WD EDIT: RadialSelectorMenuBUI -> AttachedRadialSelectorMenuBUI
   - type: ActivatableUI
     key: enum.RadialSelectorUiKey.Key
     inHandsOnly: true
@@ -215,7 +215,7 @@
   - type: UserInterface
     interfaces:
       enum.RadialSelectorUiKey.Key:
-        type: RadialSelectorMenuBUI
+        type: AttachedRadialSelectorMenuBUI # WD EDIT: RadialSelectorMenuBUI -> AttachedRadialSelectorMenuBUI
   - type: ActivatableUI
     key: enum.RadialSelectorUiKey.Key
     inHandsOnly: true
@@ -300,7 +300,7 @@
   - type: UserInterface
     interfaces:
       enum.RadialSelectorUiKey.Key:
-        type: RadialSelectorMenuBUI
+        type: AttachedRadialSelectorMenuBUI # WD EDIT: RadialSelectorMenuBUI -> AttachedRadialSelectorMenuBUI
   - type: ActivatableUI
     key: enum.RadialSelectorUiKey.Key
     inHandsOnly: true
@@ -365,7 +365,7 @@
     - type: UserInterface
       interfaces:
         enum.RadialSelectorUiKey.Key:
-          type: RadialSelectorMenuBUI
+          type: AttachedRadialSelectorMenuBUI
     - type: ActivatableUI
       key: enum.RadialSelectorUiKey.Key
       inHandsOnly: true
