+++ Resources/Prototypes/WhiteDream/Entities/Objects/Cult/constructs.yml
@@ -17,7 +17,7 @@
   - type: UserInterface
     interfaces:
       enum.RadialSelectorUiKey.Key:
-        type: RadialSelectorMenuBUI
+        type: TrackedRadialSelectorMenuBUI
   - type: ContainerContainer
     containers:
       Shard: !type:ContainerSlot
