+++ Resources/Prototypes/_EE/Entities/Mobs/Species/biosynth.yml
@@ -29,7 +29,7 @@
         type: InstrumentBoundUserInterface
         requireInputValidation: false
       enum.RadialSelectorUiKey.Key:
-        type: RadialSelectorMenuBUI
+        type: TrackedRadialSelectorMenuBUI # WD EDIT: RadialSelectorMenuBUI -> TrackedRadialSelectorMenuBUI
       enum.ListViewSelectorUiKey.Key:
         type: ListViewSelectorBUI
       enum.SurgeryUIKey.Key:
